Full rewrite of game state equivalence (and session file encoding).
Previously, the game state was represented in the redo session as an
array of 52 bytes, each byte providing the absolute position of one
card. While simple enough to understand, this occasionally meant that
two game states would be considered inequivalent if, say, two cards in
the reserves exchanged places. If these two states were part of
solutions that could be pieced together form a new, shorter solution,
it was up to the user to discover this fact, and manually execute the
translated moves in order to realize the solution.

This change replaces the encoding of the game state, and of moves,
within the redo session. Instead of storing positions in the state
array, each element identifies the card that it rests upon. (A tableau
column thus can be traced through the state data from top to bottom,
for example.) The bottommost card of a tableau column stores a special
value representing the empty location in its element -- as do cards in
the reserve, though not the same value. Thus, no information is stored
in the state data as to which tableau place, or reserve place, a card
is located at. (Foundation piles, obviously, still encode their
precise location.) With this encoding, two game states that differ
only in the ordering of tableau columns and/or reserve cards will be
identified as equivalent.

To avoid having tableau columns bounce around unpredicatably during
game play, the inplay array is appended to the state data. It is extra
data that is not considered when determining equivalence, but is
nonetheless preserved during a session.

Moves are also encoded differently in the redo session, so that they
can be applied consistently when copied to a new branch. Instead of
encoding the move's source place, moves in the redo session encode the
card being moved (plus one bit to indicate a second-choice
destination, as before). Changing the move encoding means changing the
encoding of the session files: a backwards incompatibility, but at
this point I strongly suspect my user base could be numbered on one
hand, if not one finger, so it is not considered worthwhile for the
code to continue to support the older format.

As a happy side effect of this change, the code to apply moves and the
code to render the layout both became simpler. In fact, almost the
only code to become more complicated was the code to extract a
solution (i.e. the string of letter moves). Given the usage frequency
of these three functions, this is an excellent trade-off to have made.