Replace the gzipped bitmap resources with PNGs.
Originally the motivation for using gzipped bitmaps as the program's
internal image format was that bitmaps are the only image file format
that SDL uses natively. Using it allowed the program to avoid taking
another library dependency. This was somewhat defeated by requiring
zlib to decompressing the images, but the gzipped bitmaps were as
small as (if not smaller than) the equivalent PNGs. However, there
were problems with tools that made it difficult to use bitmaps with an
alpha channel, and so it made sense to keep the files as PNGs and only
translate them into compressed bitmaps as part of the build process.
This was done with a python script using PIL.

However, issues with python versioning, and the fact that PIL is not
always present, has made this a less attractive choice. Faced with
issues trying to set up a build-only dependency on a python library,
it makes more sense at this point to just have the program itself take
a direct dependency on libpng. And the resulting code is smaller than
the zlib code that it obviates a need for.

Also in this change: some of the image files have been compressed
using pngcrush, since they now are part of the resulting binary.