Add images.png and its attendant header files to the distribution.
These are generated file, and thus generally should not be
part of the distribution but should be created as part of the build
process. However, the generated files are small, and including them in
the distribution avoids a build-time dependency on PIL, which turns
out is NOT part of the standard python installation everywhere.

The python script and its inputs will still be part of the
distribution, so users will still be able to edit images.png by
editing the source images -- the only difference is making that step
optional.