Rename "configuration" to "game" everywhere, and replace mkconfigs.py.
Turned out to be confusing to refer to the data describing the initial
deck ordering as a "configuration", and the files with config in the
name kept conflicting with build configuration files. So
configurations are now called games, or game IDs, or game data. This
necessitated a massive change to variables and comment text.

In addition, the mkconfigs.py script, which built the binary data file
containing the decks, has been replaced by mkgames.c. It makes sense
to have this file built by an simple script written in a interpreted
language like python ... but it's not worth having it be the only
reason for the build process requiring python to be installed.