Break images.png into two parts: button icons and alert images.
This separation is done because the two types of images have two
different use cases. The button icons need to be loaded as surfaces
and are not useful once all of the buttons are created, while the
alert images are only useful as textures. Thus it makes sense to have
a separate PNG for each.

This change also removes some unused images, and the old Windows icon.
A new icon is coming soon, hopefully.