Better distribution building process.
A new dists subdirectory holds the scripts and input files used to
build the various distributions that are provided. Some of the
distributions are rather complex, and it makes more sense to give this
stuff its own place to live instead of having it floating around the
top-level directory.

This commit also adds a distribution type for making a windows
installer program via the nsis compiler.