Dump the C mkgames tool and just inclue the binary gamedata file.
Use of a compiled tool as part of the build process turned out to be
problematic when cross-compiling. Since the binary gamedata file is no
larger than some of the PNGs, the solution was to just include it in
the source. (The text file and python script are still included as
well, but like the built PNGs they are not a required step in the
build process.)